Remove translation for language name "Moldavian" as it's simply Romanian
According to Wikipedia, the current linguistic consensus, as well as
current political practice in Moldova, is that "Moldavian" is a synonym
for Romanian. Splitting the language is a historic political artifact.

Moldavian could be written using Cyrillic letters but in practice, all
app translations were written in the Latin alphabet, so there is no need
to keep both translations around.